% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{database_query}
\alias{database_query}
\title{Query Notion database}
\usage{
database_query(id = getOption("id"), full = TRUE)
}
\arguments{
\item{id}{Notion database id}

\item{full}{full database or head 100 content}
}
\description{
Query Notion database
}
