% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_cover_update}
\alias{graph_cover_update}
\title{Graph for xiaohongshu cover content}
\usage{
graph_cover_update(data, input = "~/Downloads/daily_report/cover.png", output)
}
\arguments{
\item{data}{input data frame}

\item{input}{raw cover graph}

\item{output}{updated cover graph output dir}
}
\description{
update cover content
}
