% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_main_single}
\alias{graph_main_single}
\title{Graph for xiaohongshu main content}
\usage{
graph_main_single(data, file = NULL)
}
\arguments{
\item{data}{input data frame}

\item{file}{output file}
}
\description{
single unit of main content
}
