% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_main_topic}
\alias{graph_main_topic}
\title{Graph for xiaohongshu main content by topic}
\usage{
graph_main_topic(data, output, pages = NULL)
}
\arguments{
\item{data}{input data frame}

\item{output}{output dir}

\item{pages}{output pages number}
}
\description{
combine main content by topic
}
