% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_timeline}
\alias{graph_timeline}
\title{Timeline graph}
\usage{
graph_timeline(tab, output = NULL)
}
\arguments{
\item{tab}{input table}

\item{output}{output dir}
}
\description{
timeline graph separeted by topic and sort by start time
}
