% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{id_setup_global}
\alias{id_setup_global}
\title{Setup ID}
\usage{
id_setup_global(id)
}
\arguments{
\item{id}{a hash string of notion page or database id}
}
\description{
setup page or database id of notion as global variable
}
