% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{is.notion_date}
\alias{is.notion_date}
\title{jadge object type is notion_date}
\usage{
is.notion_date(object)
}
\arguments{
\item{object}{input data object}
}
\description{
if type is notion_date, return TRUE
}
