% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{is.notion_multi_select}
\alias{is.notion_multi_select}
\title{jadge object type is notion_multi_select}
\usage{
is.notion_multi_select(object)
}
\arguments{
\item{object}{input data object}
}
\description{
if type is notion_multi_select, return TRUE
}
