% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{is.notion_rich_text}
\alias{is.notion_rich_text}
\title{jadge object type is notion_rich_text}
\usage{
is.notion_rich_text(object)
}
\arguments{
\item{object}{input data object}
}
\description{
if type is notion_rich_text, return TRUE
}
