% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{is.notion_title}
\alias{is.notion_title}
\title{jadge object type is notion_title}
\usage{
is.notion_title(object)
}
\arguments{
\item{object}{input data object}
}
\description{
if type is notion_title, return TRUE
}
