% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{notion_data-class}
\alias{notion_data-class}
\alias{notion_data}
\title{The notion_data Class}
\description{
The notion_data object is an intermediate storage container used internally throughout the integration procedure to hold bits of data that are useful downstream.
}
\section{Slots}{

\describe{
\item{\code{results}}{main result database}

\item{\code{object}}{list or others}

\item{\code{next_cursor}}{hash id}

\item{\code{has_more}}{more than query limited, True or False}

\item{\code{type}}{page or others}

\item{\code{page}}{list() or others}
}}

\concept{object}
