% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_database_table,list-method}
\alias{notion_database_table,list-method}
\title{Simplify notion_data}
\usage{
\S4method{notion_database_table}{list}(object)
}
\arguments{
\item{object}{notion_data object}
}
\description{
for simplifing database
}
