% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_database_table}
\alias{notion_database_table}
\title{Simplify notion_data generic method}
\usage{
notion_database_table(object)
}
\arguments{
\item{object}{notion_data object}
}
\description{
for filtering records in date region
}
