% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{notion_date-class}
\alias{notion_date-class}
\alias{notion_date}
\title{The notion_date Class}
\description{
In Notion database, there is a type of result, date.
}
\section{Slots}{

\describe{
\item{\code{id}}{data element id}

\item{\code{type}}{data type}

\item{\code{start}}{start time}

\item{\code{end}}{end time}

\item{\code{time_zone}}{time zone}
}}

\concept{object}
