% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_filter,notion_date-method}
\alias{notion_filter,notion_date-method}
\title{filter notion_date}
\usage{
\S4method{notion_filter}{notion_date}(object, min = today(), max = today())
}
\arguments{
\item{object}{notion_date object}

\item{min}{min date}

\item{max}{max date}
}
\description{
for filtering records in date region
}
