% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_filter}
\alias{notion_filter}
\title{filter notion_date generic method}
\usage{
notion_filter(object, min, max)
}
\arguments{
\item{object}{notion_date object}

\item{min}{min date}

\item{max}{max date}
}
\description{
for filtering records in date region
}
