% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{notion_multi_select-class}
\alias{notion_multi_select-class}
\alias{notion_multi_select}
\title{The notion_multi_select Class}
\description{
In Notion database, there is a type of result, multi select.
}
\section{Slots}{

\describe{
\item{\code{id}}{data element id}

\item{\code{type}}{end time}

\item{\code{data}}{data type}
}}

\concept{object}
