% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{notion_rich_text-class}
\alias{notion_rich_text-class}
\alias{notion_rich_text}
\title{The notion_rich_text Class}
\description{
In Notion database, there is a type of result, title.
}
\section{Slots}{

\describe{
\item{\code{id}}{data element id}

\item{\code{type}}{data type}

\item{\code{text}}{data content}

\item{\code{link}}{url link}

\item{\code{annotations}}{style properties}

\item{\code{plain_text}}{data content without properties}

\item{\code{href}}{url link}
}}

\concept{object}
