% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_simplify,notion_date-method}
\alias{notion_simplify,notion_date-method}
\title{Simplify notion_date}
\usage{
\S4method{notion_simplify}{notion_date}(object)
}
\arguments{
\item{object}{notion_date object}
}
\description{
for simplifing notion_date object to simple type output
}
