% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_simplify,notion_multi_select-method}
\alias{notion_simplify,notion_multi_select-method}
\title{Simplify notion_multi_select}
\usage{
\S4method{notion_simplify}{notion_multi_select}(object)
}
\arguments{
\item{object}{notion_multi_select object}
}
\description{
for simplifing notion_multi_select object to simple type output
}
