% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_simplify,notion_rich_text-method}
\alias{notion_simplify,notion_rich_text-method}
\title{Simplify notion_rich_text}
\usage{
\S4method{notion_simplify}{notion_rich_text}(object)
}
\arguments{
\item{object}{notion_rich_text object}
}
\description{
for simplifing notion_rich_text object to simple type output
}
