% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_simplify,notion_title-method}
\alias{notion_simplify,notion_title-method}
\title{Simplify notion_title}
\usage{
\S4method{notion_simplify}{notion_title}(object)
}
\arguments{
\item{object}{notion_title object}
}
\description{
for simplifing notion_title object to simple type output
}
