% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{notion_simplify}
\alias{notion_simplify}
\title{simplify notion_* object generic method}
\usage{
notion_simplify(object)
}
\arguments{
\item{object}{notion_* object}
}
\description{
for simplifing different notion object to same type output
}
