% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object_create_auto}
\alias{object_create_auto}
\title{Create notion_object automatically}
\usage{
object_create_auto(list)
}
\arguments{
\item{list}{input list}
}
\description{
create notion_object based on type
}
