% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object_create_date}
\alias{object_create_date}
\title{Create notion_date Object}
\usage{
object_create_date(list)
}
\arguments{
\item{list}{input data in list format}
}
\value{
Returns a notion_date object
}
\description{
create a new object for date data
}
\concept{object}
