% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object_create_multi_select}
\alias{object_create_multi_select}
\title{Create notion_multi_select Object}
\usage{
object_create_multi_select(list)
}
\arguments{
\item{list}{input data in list format}
}
\value{
Returns a notion_multi_select object
}
\description{
create a new object for multi select data
}
\concept{object}
