% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object_create_rich_text}
\alias{object_create_rich_text}
\title{Create notion_rich_text Object}
\usage{
object_create_rich_text(list)
}
\arguments{
\item{list}{input data in list format}
}
\value{
Returns a notion_rich_text object
}
\description{
create a new object for rich_text data
}
\concept{object}
