% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object_create_title}
\alias{object_create_title}
\title{Create notion_title Object}
\usage{
object_create_title(list)
}
\arguments{
\item{list}{input data in list format}
}
\value{
Returns a notion_title object
}
\description{
create a new object for title data
}
\concept{object}
