% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_notion_xiaohongshu}
\alias{pipeline_notion_xiaohongshu}
\title{Pipeline xiaohongshu
automatic workflow to generate content for social media based on Notion database}
\usage{
pipeline_notion_xiaohongshu(dir)
}
\arguments{
\item{dir}{output dir}
}
\description{
Pipeline xiaohongshu
automatic workflow to generate content for social media based on Notion database
}
