% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{replace.values}
\alias{replace.values}
\title{Data frame replace elements}
\usage{
replace.values(search, replace, x)
}
\arguments{
\item{search}{old value}

\item{replace}{new value}

\item{x}{input data}
}
\description{
using same length vector to replace data frame elements. from https://stackoverflow.com/questions/11810605/replace-contents-of-factor-column-in-r-dataframe
}
