% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{show,notion_date-method}
\alias{show,notion_date-method}
\title{show notion_date}
\usage{
\S4method{show}{notion_date}(object)
}
\arguments{
\item{object}{notion_date object}
}
\description{
show method for S4 class notion_date
}
