% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{show,notion_multi_select-method}
\alias{show,notion_multi_select-method}
\title{show notion_multi_select}
\usage{
\S4method{show}{notion_multi_select}(object)
}
\arguments{
\item{object}{notion_multi_select object}
}
\description{
show method for S4 class notion_multi_select
}
