% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{show,notion_rich_text-method}
\alias{show,notion_rich_text-method}
\title{show notion_rich_text}
\usage{
\S4method{show}{notion_rich_text}(object)
}
\arguments{
\item{object}{notion_rich_text object}
}
\description{
show method for S4 class notion_rich_text
}
