% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{show,notion_title-method}
\alias{show,notion_title-method}
\title{show notion_title}
\usage{
\S4method{show}{notion_title}(object)
}
\arguments{
\item{object}{notion_title object}
}
\description{
show method for S4 class notion_title
}
