% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_build_graph}
\alias{table_build_graph}
\title{Create main table for xiaohongshu graph content}
\usage{
table_build_graph(tab, index)
}
\arguments{
\item{tab}{input data}

\item{index}{table row index}
}
\description{
based on the database table to generate a frame of graph
}
