% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_mutate_time}
\alias{table_mutate_time}
\title{Mutate new column of start and end time}
\usage{
table_mutate_time(tab)
}
\arguments{
\item{tab}{input data}
}
\description{
based on the database table time column of character format to generate start and end time in POSIXct format
}
