% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_to_report}
\alias{table_to_report}
\title{Table to Daily Report}
\usage{
table_to_report(file)
}
\arguments{
\item{file}{input file in csv format}
}
\description{
Generate daily report content based on Notion csv format database
}
