% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{token_setup_global}
\alias{token_setup_global}
\title{Setup API Token}
\usage{
token_setup_global(
  file = system.file("extdata", "token.txt", package = "ttnotion")
)
}
\arguments{
\item{file}{token file}
}
\description{
setup global variable of token for notion api
}
