% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_cover_update}
\alias{graph_cover_update}
\title{Graph for xiaohongshu cover content}
\usage{
graph_cover_update(
  data,
  input = system.file("extdata", "cover.png", package = "ttnotion"),
  output
)
}
\arguments{
\item{data}{input data frame}

\item{input}{raw cover graph}

\item{output}{updated cover graph output dir}
}
\description{
update cover content
}
