% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{object_to_report}
\alias{object_to_report}
\title{Notion database object to Daily Report}
\usage{
object_to_report(object, dir)
}
\arguments{
\item{object}{input simplified notion database object}

\item{dir}{output dir}
}
\description{
Generate daily report content based on Notion database object
}
